requirements.txt
django~=3.2.6
djangorestframework~=3.12.4
django-cors-headers~=3.8.0
flake8~=3.9.2
cryptography~=3.4.8
pyjwt~=2.1.0
python-dotenv~=0.19.0